% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-aging.R
\name{aging}
\alias{aging}
\title{Aging Module}
\usage{
aging(dat, at)
}
\arguments{
\item{dat}{Main data object of class `netsim_dat` containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns `dat` after updating the nodal attributes `age` and
`sqrt.age`.
}
\description{
Module for aging over time for active nodes in the population.
}
