% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_attributes.R
\name{get_default_attrs}
\alias{get_default_attrs}
\title{Get the list the attributes used by the model with their default value}
\usage{
get_default_attrs(dat)
}
\value{
A named list of all the attributes with default values.

TODO: document all the attributes
}
\description{
Get the list the attributes used by the model with their default value
}
\details{
This list must be exhaustive. All attributes should get a default value (even
NA) here.
}
