% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_attributes.R
\name{make_computed_attrs}
\alias{make_computed_attrs}
\title{Update the attributes requiring computation for new nodes
This function returns the `dat` object with updated attributes for the new
nodes.}
\usage{
make_computed_attrs(dat, n_new, post_init)
}
\description{
Update the attributes requiring computation for new nodes
This function returns the `dat` object with updated attributes for the new
nodes.
}
