% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-dx.R,
%   R/mod-progress.R, R/mod-vax.R
\name{moduleset-common}
\alias{moduleset-common}
\alias{dx_covid_corporate}
\alias{progress_covid_corporate}
\alias{vax_covid_corporate}
\title{EpiModel Common Module Set Across COVID Models}
\usage{
dx_covid_corporate(dat, at)

progress_covid_corporate(dat, at)

vax_covid_corporate(dat, at)
}
\arguments{
\item{dat}{Main `netsim_dat` class data object of network models.}

\item{at}{Current time step.}
}
\description{
This set of functions is associated with all EpiModel studies of
             COVID.

dx.rate.sympt <- get_param(dat, "dx.rate.sympt")
  if (length(dx.rate.sympt) > 1) {
    dx.rate.sympt <- dx.rate.sympt[at]
  }
  dx.rate.other <- get_param(dat, "dx.rate.other")
  if (length(dx.rate.other) > 1) {
    dx.rate.other <- dx.rate.other[at]
  }
  allow.rescreen <- get_param(dat, "allow.rescreen")
  pcr.sens <- get_param(dat, "pcr.sens")
}
\details{

}
