% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R
\name{moduleset-contacttrace}
\alias{moduleset-contacttrace}
\title{EpiModel Module Set for Contact Tracing Model}
\arguments{
\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}

\item{dat}{Master list object of network models.}

\item{at}{Current time step.}
}
\description{
This set of functions is associated with the EpiModel study of
             contact tracing for COVID.
}
