% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-isolate.R,
%   R/mod-prevalence.R
\name{moduleset-corporate}
\alias{moduleset-corporate}
\alias{isolate_covid_corporate}
\alias{prevalence_covid_corporate}
\title{EpiModel Module Set for Corporate Office Model}
\usage{
isolate_covid_corporate(dat, at)

prevalence_covid_corporate(dat, at)
}
\arguments{
\item{dat}{Main `netsim_dat` class data object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with the EpiModel study of
             COVID in corporate office environments.
}
