% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-demog.R,
%   R/mod-infection.R, R/mod-init.R, R/mod-prevalence.R, R/mod-simnet.R
\name{moduleset-netjail}
\alias{moduleset-netjail}
\alias{deaths_covid_netjail}
\alias{covid_release_netjail}
\alias{arrival_covid_netjail}
\alias{infect_covid_netjail}
\alias{init_covid_netjail}
\alias{prevalence_covid_netjail}
\alias{resim_nets_covid_netjail}
\title{EpiModel Module Set for Fulton County Jail Model}
\usage{
deaths_covid_netjail(dat, at)

covid_release_netjail(dat, at)

arrival_covid_netjail(dat, at)

infect_covid_netjail(dat, at)

init_covid_netjail(x, param, init, control, s)

prevalence_covid_netjail(dat, at)

resim_nets_covid_netjail(dat, at)
}
\arguments{
\item{dat}{Master list object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with the EpiModel study of
             COVID in a jail/prison setting.
}
