% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-demog.R,
%   R/mod-infection.R, R/mod-init.R, R/mod-prevalence.R, R/mod-progress.R,
%   R/mod-simnet.R, R/mod-vax.R
\name{moduleset-vaxDecisions}
\alias{moduleset-vaxDecisions}
\alias{deaths_covid_vax_decisions}
\alias{arrival_covid_vax_decisions}
\alias{infect_covid_vax_decisions}
\alias{init_covid_vax_decisions}
\alias{prevalence_covid_vax_decisions}
\alias{progress_covid_vax_decisions}
\alias{resim_nets_covid_vax_decisions}
\alias{vax_covid_vax_decisions}
\title{EpiModel Module Set for Vaccination Decision-Making Model}
\usage{
deaths_covid_vax_decisions(dat, at)

arrival_covid_vax_decisions(dat, at)

infect_covid_vax_decisions(dat, at)

init_covid_vax_decisions(x, param, init, control, s)

prevalence_covid_vax_decisions(dat, at)

progress_covid_vax_decisions(dat, at)

resim_nets_covid_vax_decisions(dat, at)

vax_covid_vax_decisions(dat, at)
}
\arguments{
\item{dat}{Main `netsim_dat` class data object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with the EpiModel study of
             vaccination decision making.
}
