% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R
\docType{package}
\name{EpiModelCOVID-package}
\alias{EpiModelCOVID-package}
\alias{EpiModelCOVID}
\title{Model Extensions to EpiModel for the Network-Based Modeling of COVID-19}
\description{
\tabular{ll}{
   Package: \tab EpiModelCOVID\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2020-05-17\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\keyword{package}
