% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_attributes.R
\name{get_default_attrs}
\alias{get_default_attrs}
\title{Get the list the attributes used by the model with their default value}
\usage{
get_default_attrs(dat)
}
\description{
Get the list the attributes used by the model with their default value
}
