% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_attributes.R
\name{make_computed_attrs}
\alias{make_computed_attrs}
\title{Update the attributes requiring computation for new nodes}
\usage{
make_computed_attrs(dat, n_new, post_init)
}
\arguments{
\item{n_new}{The number of new nodes to update}

\item{post_init}{logical flag, TRUE if not called by the initialization
module.}
}
\value{
This function returns the `dat` object with updated attributes for the new
nodes.
}
\description{
Update the attributes requiring computation for new nodes
}
\details{
This function takes care of all the attributes requiring some computations.
This includes the random assignment for race or the calculation of the
age groups. New attributes that need special assignment should be set here
AS WELL AS in the `get_default_attrs` function.
}
