% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-dx.R,
%   R/mod-progress.R
\name{moduleset-common}
\alias{moduleset-common}
\alias{dx_covid_corporate}
\alias{progress_general}
\title{EpiModel Common Module Set Across COVID Models}
\usage{
dx_covid_corporate(dat, at)

progress_general(dat, at)
}
\arguments{
\item{dat}{Main `netsim_dat` class data object of network models.}

\item{at}{Current time step.}
}
\description{
This set of functions is associated with all EpiModel studies of
             COVID.
}
