% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-demog.R,
%   R/mod-dx.R, R/mod-intervention.R, R/mod-progress.R, R/mod-vax.R
\name{moduleset-common}
\alias{moduleset-common}
\alias{aging_covid}
\alias{dx_covid}
\alias{contact_trace_covid}
\alias{progress_covid}
\alias{vax_covid}
\title{EpiModel Common Module Set Across COVID Models}
\usage{
aging_covid(dat, at)

dx_covid(dat, at)

contact_trace_covid(dat, at)

progress_covid(dat, at)

vax_covid(dat, at)
}
\arguments{
\item{dat}{Master list object of network models.}

\item{at}{Current time step.}
}
\description{
This set of functions is associated with all EpiModel studies of
             COVID.
}
