% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-demog.R,
%   R/mod-dx.R, R/mod-progress.R
\name{moduleset-common}
\alias{moduleset-common}
\alias{aging_covid}
\alias{dx_covid}
\alias{progress_covid}
\title{EpiModel Common Module Set Across COVID Models}
\usage{
aging_covid(dat, at)

dx_covid(dat, at)

progress_covid(dat, at)
}
\arguments{
\item{dat}{Master list object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with all EpiModel studies of
             COVID.
}
