% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-demog.R,
%   R/mod-dx.R, R/mod-infection.R, R/mod-init.R, R/mod-prevalence.R,
%   R/mod-progress.R, R/mod-simnet.R
\name{moduleset-contacttrace}
\alias{moduleset-contacttrace}
\alias{deaths_covid_contacttrace}
\alias{arrival_covid_contacttrace}
\alias{dx_covid_contacttrace}
\alias{infect_covid_contacttrace}
\alias{init_covid_contacttrace}
\alias{prevalence_covid_contacttrace}
\alias{progress_covid_contacttrace}
\alias{resim_nets_covid_contacttrace}
\title{EpiModel Module Set for Contact Tracing Model}
\usage{
deaths_covid_contacttrace(dat, at)

arrival_covid_contacttrace(dat, at)

dx_covid_contacttrace(dat, at)

infect_covid_contacttrace(dat, at)

init_covid_contacttrace(x, param, init, control, s)

prevalence_covid_contacttrace(dat, at)

progress_covid_contacttrace(dat, at)

resim_nets_covid_contacttrace(dat, at)
}
\arguments{
\item{dat}{Master list object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with the EpiModel study of
             contact tracing for COVID.
}
