% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-arrival.R,
%   R/mod-departure.R, R/mod-infection.R, R/mod-init.R, R/mod-netdegree.R,
%   R/mod-vax.R
\name{moduleset-gmc19}
\alias{moduleset-gmc19}
\alias{arrival}
\alias{deaths_covid_gmc19}
\alias{infect_general}
\alias{init_gmc19}
\alias{netdegree}
\alias{vax_general}
\alias{compute_ve}
\alias{vax_age_group_for}
\title{EpiModel Module Set for GMC-19 India Model}
\usage{
arrival(dat, at)

deaths_covid_gmc19(dat, at)

infect_general(dat, at)

init_gmc19(x, param, init, control, s)

netdegree(dat, at)

vax_general(dat, at)

compute_ve(
  at,
  ids,
  vax,
  vax.age.group,
  vax1Time = NULL,
  vax2Time = NULL,
  vax3Time = NULL,
  vax.schedule,
  outcome = c("infect", "clinical", "hosp")
)

vax_age_group_for(dat)
}
\arguments{
\item{dat}{Main `netsim_dat` class data object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with the EpiModel study of
             COVID in the India GMC-19 model.
}
