% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R
\name{moduleset-ship}
\alias{moduleset-ship}
\title{EpiModel Module Set for COVID Cruise Ship Model}
\arguments{
\item{x}{An \code{EpiModel} object of class `netest`.}

\item{param}{An \code{EpiModel} object of class `param.net`.}

\item{init}{An \code{EpiModel} object of class `init.net`.}

\item{control}{An \code{EpiModel} object of class `control.net`.}

\item{s}{Simulation number, used for restarting dependent simulations.}

\item{dat}{Main `netsim_dat` class data object of network models.}

\item{at}{Current time step.}
}
\description{
This set of functions is associated with the EpiModel study of
             COVID on cruise ship environments.
}
