% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelCOVID-package.R, R/mod-demog.R,
%   R/mod-dx.R, R/mod-infection.R, R/mod-init.R, R/mod-prevalence.R,
%   R/mod-progress.R, R/mod-simnet.R
\name{moduleset-ship}
\alias{moduleset-ship}
\alias{aging_covid_ship}
\alias{deaths_covid_ship}
\alias{offload_covid_ship}
\alias{dx_covid_ship}
\alias{infect_covid_ship}
\alias{discord_edgelist_covid_ship}
\alias{init_covid_ship}
\alias{init_status_covid_ship}
\alias{prevalence_covid_ship}
\alias{progress_covid_ship}
\alias{resim_nets_covid_ship}
\alias{edges_correct_covid}
\alias{calc_nwstats_covid}
\title{EpiModel Module Set for COVID Cruise Ship Model}
\usage{
aging_covid_ship(dat, at)

deaths_covid_ship(dat, at)

offload_covid_ship(dat, at)

dx_covid_ship(dat, at)

infect_covid_ship(dat, at)

discord_edgelist_covid_ship(dat, nw = 1)

init_covid_ship(x, param, init, control, s)

init_status_covid_ship(dat)

prevalence_covid_ship(dat, at)

progress_covid_ship(dat, at)

resim_nets_covid_ship(dat, at)

edges_correct_covid(dat, at)

calc_nwstats_covid(dat, at)
}
\arguments{
\item{dat}{Master list object of network models.}

\item{at}{Current time step.}

\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param_msm}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init_msm}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control_msm}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This set of functions is associated with the EpiModel study of
             COVID on cruise ship environments.
}
